+++ spec/models/user_activity_spec.rb
@@ -33,6 +33,41 @@ describe UserActivity do
         ]
       )
     end
+    context 'with all fields' do
+      before do
+        team.activity_fields = ['All']
+      end
+      it 'to_slack' do
+        expect(activity.to_slack).to eq(
+          attachments: [
+            {
+              fallback: "#{activity.name} via #{activity.user.slack_mention}, 14.01mi 2h6m26s 9m02s/mi",
+              title: activity.name,
+              title_link: "https://www.strava.com/activities/#{activity.strava_id}",
+              text: "<@#{activity.user.user_name}> on Tuesday, February 20, 2018 at 10:02 AM",
+              image_url: "https://slava.playplay.io/api/maps/#{activity.map.id}.png",
+              fields: [
+                { title: 'Type', value: 'Run 🏃', short: true },
+                { title: 'Distance', value: '14.01mi', short: true },
+                { title: 'Moving Time', value: '2h6m26s', short: true },
+                { title: 'Elapsed Time', value: '2h8m6s', short: true },
+                { title: 'Pace', value: '9m02s/mi', short: true },
+                { title: 'Speed', value: '6.6mph', short: true },
+                { title: 'Elevation', value: '475.4ft', short: true },
+                { title: 'Max Speed', value: '20.8mph', short: true },
+                { title: 'Heart Rate', value: '140.3bpm', short: true },
+                { title: 'Max Heart Rate', value: '178.0bpm', short: true },
+                { title: 'PR Count', value: '3', short: true },
+                { title: 'Calories', value: '870.2', short: true }
+              ],
+              author_name: user.athlete.name,
+              author_link: user.athlete.strava_url,
+              author_icon: user.athlete.profile_medium
+            }
+          ]
+        )
+      end
+    end
     context 'without an athlete' do
       before do
         user.athlete.destroy